move _to_dense and _non_zero to NamedArray
